Enhance showcase editor in example
