+++ @remirror/editor-social/package.json
@@ -19,7 +19,6 @@
     "@emotion/styled": "^10.0.17",
     "@remirror/core": "^0.7.3",
     "@remirror/core-extensions": "^0.7.3",
-    "@remirror/extension-code-block": "^0.7.3",
     "@remirror/extension-emoji": "^0.7.3",
     "@remirror/extension-enhanced-link": "^0.7.3",
     "@remirror/extension-mention": "^0.7.3",
@@ -28,8 +27,7 @@
     "@remirror/ui": "^0.7.3",
     "@remirror/ui-buttons": "^0.7.3",
     "@remirror/ui-icons": "^0.7.3",
-    "multishift": "^0.7.3",
-    "refractor": "^2.10.0"
+    "multishift": "^0.7.3"
   },
   "peerDependencies": {
     "@emotion/core": "^10",
