+++ @remirror/editor-social/tsconfig.prod.json
@@ -28,9 +28,6 @@
       "path": "../core-extensions/tsconfig.prod.json"
     },
     {
-      "path": "../extension-code-block/tsconfig.prod.json"
-    },
-    {
       "path": "../extension-emoji/tsconfig.prod.json"
     },
     {
